sandbox: cleanup plugin and ensure options and netns are exposed
